Separation of SourceLike, fwi operators in acoustic
